Rename package db to store
Its upcoming implementation is not related to db.
