+++ pkg/gateway/provider/config.go
@@ -5,14 +5,14 @@ import (
 
 	"github.com/skygeario/skygear-server/pkg/core/config"
 	coreMiddleware "github.com/skygeario/skygear-server/pkg/core/middleware"
-	"github.com/skygeario/skygear-server/pkg/gateway/db"
 	"github.com/skygeario/skygear-server/pkg/gateway/model"
+	"github.com/skygeario/skygear-server/pkg/gateway/store"
 )
 
 // GatewayTenantConfigurationProvider provide tenlent config from request
 type GatewayTenantConfigurationProvider struct {
 	coreMiddleware.ConfigurationProvider
-	Store db.GatewayStore
+	Store store.GatewayStore
 }
 
 // ProvideConfig function query the tenant config from db by request
